authorization for common and backoffice

+++ TestingAndCalibrationLabs/TestingAndCalibrationLabs.Business/Services/BackOffice/UiControlTypeService.cs
@@ -1,4 +1,6 @@
-﻿using System.Collections.Generic;
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Http;
+using System.Collections.Generic;
 using TestingAndCalibrationLabs.Business.Common;
 using TestingAndCalibrationLabs.Business.Core.Interfaces;
 using TestingAndCalibrationLabs.Business.Core.Model;
@@ -12,11 +14,14 @@ namespace TestingAndCalibrationLabs.Business.Services
     public class UiControlTypeService : IUiControlTypeService
     {
         private readonly IGenericRepository<UiControlTypeModel> _genericRepository;
-        public UiControlTypeService( IGenericRepository<UiControlTypeModel> genericRepository)
+        private readonly IAuthorizationService _authorizationService;
+        private readonly IHttpContextAccessor _httpContextAccessor;
+        public UiControlTypeService( IGenericRepository<UiControlTypeModel> genericRepository,IAuthorizationService authorizationService,IHttpContextAccessor httpContextAccessor)
         {
             _genericRepository = genericRepository;
+            _authorizationService = authorizationService;
+            _httpContextAccessor = httpContextAccessor;
         }
-
         #region Public methods
         /// <summary>
         /// Get Record By Id For Ui Control Type
@@ -25,7 +30,9 @@ namespace TestingAndCalibrationLabs.Business.Services
         /// <returns></returns>
         public UiControlTypeModel GetById(int id)
         {
-            return _genericRepository.Get(id);
+            if (_authorizationService.AuthorizeAsync(_httpContextAccessor.HttpContext.User, new UiControlTypeModel(), Operations.Read).Result.Succeeded)
+                return _genericRepository.Get(id);
+            return null;
         }
         /// <summary>
         /// Get All Record For Ui Control Type
@@ -33,7 +40,9 @@ namespace TestingAndCalibrationLabs.Business.Services
         /// <returns></returns>
         public List<UiControlTypeModel> Get()
         {
-            return _genericRepository.Get();
+            if (_authorizationService.AuthorizeAsync(_httpContextAccessor.HttpContext.User, new UiControlTypeModel(), Operations.Read).Result.Succeeded)
+                return _genericRepository.Get();
+            return null;
         }
         /// <summary>
         /// Edit Record From Ui Control Type
@@ -43,19 +52,26 @@ namespace TestingAndCalibrationLabs.Business.Services
         /// <returns></returns>
         public RequestResult<int> Update( UiControlTypeModel uiControlTypeModel)
         {
-            _genericRepository.Update(uiControlTypeModel);
-            return new RequestResult<int>(1);
+            if (_authorizationService.AuthorizeAsync(_httpContextAccessor.HttpContext.User, uiControlTypeModel, Operations.Update).Result.Succeeded)
+            { 
+                _genericRepository.Update(uiControlTypeModel);
+                return new RequestResult<int>(1);
+            }
+            return new RequestResult<int>(0);
         }
         /// <summary>
         /// Insert Record In Ui Control Type
         /// </summary>
         /// <param name="uiControlTypeModel"></param>
         /// <returns></returns>
-
         public RequestResult<int> Create(UiControlTypeModel uiControlTypeModel)
         {
-            _genericRepository.Insert(uiControlTypeModel);
-            return new RequestResult<int>(1);
+            if (_authorizationService.AuthorizeAsync(_httpContextAccessor.HttpContext.User, uiControlTypeModel, Operations.Create).Result.Succeeded)
+            {
+                _genericRepository.Insert(uiControlTypeModel);
+                return new RequestResult<int>(1);
+            }
+            return new RequestResult<int>(0);
         }
         /// <summary>
         /// Delete Record From Ui Control 
@@ -64,12 +80,10 @@ namespace TestingAndCalibrationLabs.Business.Services
         /// <returns></returns>
         public bool Delete(int id)
         {
-            return _genericRepository.Delete(id);
+            if (_authorizationService.AuthorizeAsync(_httpContextAccessor.HttpContext.User, new UiControlTypeModel(), Operations.Delete).Result.Succeeded)
+                return _genericRepository.Delete(id);
+            return false;
         }
-        #endregion
-
-        #region Private Methods
-
         #endregion
     }
 }
